fix: uint type

+++ include/enet/types.h
@@ -7,8 +7,8 @@
 
 typedef unsigned char enet_uint8;       /**< unsigned 8-bit type  */
 typedef unsigned short enet_uint16;     /**< unsigned 16-bit type */
-typedef unsigned int enet_uint32;      /**< unsigned 32-bit type */
-typedef unsigned int enet_uint64;      /**< unsigned 64-bit type */
+typedef unsigned int enet_uint32;       /**< unsigned 32-bit type */
+typedef unsigned long long enet_uint64; /**< unsigned 64-bit type */
 
 #endif /* __ENET_TYPES_H__ */
